refactor(Hint, Computer, User): 특정 동작 객체가 수행하도록 분리
1. Computer 클래스 추가, 정답 숫자 생성 작업 Computer 객체가 하도록 메소드 작성
2. User 클래스 추가, 숫자 입력 User 객체가 하도록 메소드 작성
3. Hint.getMessage() => Hint.printHintMessage로 변경. 객체가 출력.
